requirements.txt
streamlit>=1.32.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
pandas>=2.2.0
plotly>=5.20.0
anthropic>=0.25.0
python-dotenv>=1.0.0
